ml/requirements.txt
fastapi
uvicorn
pydantic
pandas
numpy
spacy
nltk
textblob
word2number
pillow
pymongo
fasttext
python-multipart
aiofiles
python-dotenv
pyaudio
google-generativeai
google-cloud-speech
scikit-learn
joblib
wheel
gensim
google-cloud-storage
scipy
